Add workaround to plotting
