--- doc/pdoc_template/credits.mako
@@ -1,10 +1,0 @@
-<%!
-from backtesting import __version__
-%>
-<p>
-    <a href="https://kernc.github.io/backtesting.py/"><cite>backtesting</cite> ${__version__}</a>
-    <span style="color:#ddd">&#21328;</span>
-</p>
-
-<script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.12.0/highlight.min.js"></script>
-<script>hljs.configure({languages: ['python']}); hljs.initHighlightingOnLoad()</script>
