--- doc/pdoc_template/logo.mako
@@ -1,5 +1,0 @@
-<header>
-    <a class="homelink" rel="home" title="Backtesting.py Home" href="https://kernc.github.io/backtesting.py/">
-        <img src="https://kernc.github.io/backtesting.py/logo.png" alt=""> Backtesting.py
-    </a>
-</header>
