+++ requirements.txt
@@ -1,2 +1,4 @@
-# To run example notebooks, install required and test dependencies
-.[test]
+git+ssh://git@github.com/Casaplaya/timeutil.git@v1.0.2
+numpy == 1.22.3
+pandas == 1.4.1
+bokeh == 2.4.2
